requirements.txt
python==3.10.13

matplotlib==3.8.0
networkx==3.1
pyg==2.4.0
pytorch==1.12.1